test: reduce boundary smoke theater
